Replace geoip database by local database
Fixes #11493

The MySQL had a copy of a geoip database which was queries
to estimate the position of a host on the globe.

The database was huge and querying it was slow because
MySQL is not very good with spacial queries. On top of that,
this precision wasn't really needed.

A local database is way quicker to query and provides us
with the country an IP address is assumed to be in. That
will allow us to select a few mirror servers as preferred
mirrors.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>